Separate unzip functions from Downloader
