+++ MultiplayerExtensions/HarmonyPatches/SongDownloadingPatches.cs
@@ -50,11 +50,11 @@ namespace MultiplayerExtensions.HarmonyPatches
             if (SongCore.Loader.GetLevelById(levelId) != null)
             {
                 if (LoadingLevelId != levelId)
-                    Plugin.Log?.Debug($"Level with ID '{levelId}' already exists.");
+                    Plugin.Log?.Debug($"Level with ID '{levelId}' already exists."); // Don't log if LoadLevel was called when a download finished.
                 LoadingLevelId = null;
                 return true;
             }
-            string? hash = Utilities.Utilities.LevelIdToHash(beatmapId.levelID);
+            string? hash = Utilities.Utils.LevelIdToHash(beatmapId.levelID);
             if (hash == null)
             {
                 Plugin.Log?.Info($"Could not get a hash from beatmap with LevelId {beatmapId.levelID}");
@@ -122,7 +122,7 @@ namespace MultiplayerExtensions.HarmonyPatches
         {
             if (beatmapId != null)
             {
-                string? hash = Utilities.Utilities.LevelIdToHash(beatmapId.levelID);
+                string? hash = Utilities.Utils.LevelIdToHash(beatmapId.levelID);
                 if (hash != null)
                 {
                     Plugin.Log?.Debug($"'{userId}' selected song '{hash}'.");
@@ -137,7 +137,7 @@ namespace MultiplayerExtensions.HarmonyPatches
                     BeatmapCharacteristicSO? characteristic = characteristicCollection.GetBeatmapCharacteristicBySerializedName(beatmapId.beatmapCharacteristicSerializedName);
 
                     Plugin.Log?.Debug("Setting song preview");
-                    Task<Beatmap>? beatmap = BeatSaver.Client.Hash(Utilities.Utilities.LevelIdToHash(beatmapId.levelID));
+                    Task<Beatmap>? beatmap = BeatSaver.Client.Hash(Utilities.Utils.LevelIdToHash(beatmapId.levelID));
                     beatmap.ContinueWith(r =>
                     {
                         if (r.IsCanceled)
@@ -168,7 +168,7 @@ namespace MultiplayerExtensions.HarmonyPatches
     {
         static bool Prefix(string levelId, BeatmapDifficulty beatmapDifficulty, BeatmapCharacteristicSO characteristic, LobbyPlayersDataModel __instance)
         {
-            string? hash = Utilities.Utilities.LevelIdToHash(levelId);
+            string? hash = Utilities.Utils.LevelIdToHash(levelId);
             if (hash != null)
             {
                 Plugin.Log?.Debug($"Local user selected song '{levelId}'.");
