+++ MultiplayerExtensions/Utilities/Utils.cs
@@ -9,7 +9,7 @@ using UnityEngine;
 
 namespace MultiplayerExtensions.Utilities
 {
-    public static class Utilities
+    public static class Utils
     {
         /// <summary>
         /// Logger for debugging sprite loads.
@@ -58,6 +58,46 @@ namespace MultiplayerExtensions.Utilities
             if ((hash?.Length ?? 0) == 40)
                 return hash;
             return null;
+        }
+
+        public static string GetSongDirectoryName(string? songKey, string songName, string levelAuthorName)
+        {
+            // BeatSaverDownloader's method of naming the directory.
+            string basePath;
+            string nameAuthor;
+            if (string.IsNullOrEmpty(levelAuthorName))
+                nameAuthor = songName;
+            else
+                nameAuthor = $"{songName} - {levelAuthorName}";
+            songKey = songKey?.Trim();
+            if (songKey != null && songKey.Length > 0)
+                basePath = songKey + " (" + nameAuthor + ")";
+            else
+                basePath = nameAuthor;
+            basePath = string.Concat(basePath.Trim().Split(InvalidPathChars));
+            return basePath;
+        }
+
+        private static readonly char[] _baseInvalidPathChars = new char[]
+            {
+                '<', '>', ':', '/', '\\', '|', '?', '*', '"',
+                '\u0000', '\u0001', '\u0002', '\u0003', '\u0004', '\u0005', '\u0006', '\u0007',
+                '\u0008', '\u0009', '\u000a', '\u000b', '\u000c', '\u000d', '\u000e', '\u000d',
+                '\u000f', '\u0010', '\u0011', '\u0012', '\u0013', '\u0014', '\u0015', '\u0016',
+                '\u0017', '\u0018', '\u0019', '\u001a', '\u001b', '\u001c', '\u001d', '\u001f',
+            };
+
+        private static char[]? _invalidPathChars;
+        public static char[] InvalidPathChars
+        {
+            get
+            {
+                if (_invalidPathChars == null)
+                {
+                    _invalidPathChars = _baseInvalidPathChars.Concat(Path.GetInvalidPathChars()).Distinct().ToArray();
+                }
+                return _invalidPathChars;
+            }
         }
     }
 }
